add new Spring Code

+++ src/main/java/edu/daniel/projectJPA/repositories/IAnimalesRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import edu.daniel.projectJPA.entities.Animal;
 
-public interface IAnimalesRepository extends JpaRepository <Animal, Long>{
+public interface IAnimalesRepository extends JpaRepository <Animal, Integer>{
 
 }
